refactor(OUT-1608): organize imports
