+++ src/hoc/ClientSideStateUpdate.tsx
@@ -17,8 +17,8 @@ import {
 import { setAssigneeSuggestion, setExpandedComments } from '@/redux/features/taskDetailsSlice'
 import { setTemplates } from '@/redux/features/templateSlice'
 import store from '@/redux/store'
-import { AccessibleTasksResponse, Token } from '@/types/common'
-import { TaskResponse } from '@/types/dto/tasks.dto'
+import { Token } from '@/types/common'
+import { AccessibleTasksResponse, TaskResponse } from '@/types/dto/tasks.dto'
 import { CreateViewSettingsDTO } from '@/types/dto/viewSettings.dto'
 import { WorkflowStateResponse } from '@/types/dto/workflowStates.dto'
 import { IAssigneeCombined, IAssigneeSuggestions, ITemplate } from '@/types/interfaces'
