+++ src/types/common.ts
@@ -1,7 +1,6 @@
-import { UserRole } from '@/app/api/core/types/user'
-import { AssigneeType, CommentInitiator } from '@prisma/client'
+import { UserRole } from '@api/core/types/user'
+import { CommentInitiator } from '@prisma/client'
 import { z } from 'zod'
-import { TaskResponse } from './dto/tasks.dto'
 
 export const HexColorSchema = z.string().refine((val) => /^#([0-9A-Fa-f]{3}|[0-9A-Fa-f]{6})$/.test(val), {
   message: 'Invalid hex color code',
@@ -231,5 +230,3 @@ export interface InitiatedEntity {
   initiatorId: string
   initiatorType: CommentInitiator | null
 }
-
-export type AccessibleTasksResponse = Pick<TaskResponse, 'id' | 'assigneeId' | 'assigneeType' | 'title' | 'body'>
