+++ src/types/dto/tasks.dto.ts
@@ -58,3 +58,5 @@ export const SubTaskStatusSchema = z.object({
 })
 
 export type SubTaskStatusResponse = z.infer<typeof SubTaskStatusSchema>
+
+export type AccessibleTasksResponse = Pick<TaskResponse, 'id' | 'assigneeId' | 'assigneeType' | 'title' | 'body'>
